requirements.txt
sentence-transformers
faiss-cpu
streamlit
numpy
pypdf
pyttsx3
gpt4all
